requirements.txt
ansible-base==2.10.8
ansicolors==1.1.8
apache-libcloud==3.2.0
appdirs==1.4.4
apsw==3.36.0.post1
apturl==0.5.2
argcomplete==1.8.1
argon2-cffi==21.1.0
arrow==0.14.7
art==5.6
attrs==21.2.0
autopep8==1.6.0
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==4.1.0
blinker==1.4
bottle==0.12.19
Brlapi==0.8.3
Brotli==1.0.9
cachetools==5.0.0
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
command-not-found==0.3
commonmark==0.9.1
configobj==5.0.6
cryptography==3.4.8
css-parser==1.0.7
cssselect==1.1.0
cssutils==1.0.2
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info===1.1build1
dnspython==2.1.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.4
fasteners==0.14.1
feedparser==6.0.8
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
gcalcli==4.3.0
Glances==3.2.4.2
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.1.0
gpg===1.16.0-unknown
greenlet==1.1.2
guake==3.8.5
gyp==0.1
html2text==2020.1.16
html5-parser==0.4.10
html5lib==1.1
httplib2==0.20.2
ics==0.7
idna==3.3
ifaddr==0.1.7
imgrender==0.0.4
importlib-metadata==4.6.4
influxdb==5.3.1
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
ipywidgets==6.0.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mechanize==0.4.7
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.0.3
nbclient==0.5.6
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netaddr==0.8.0
netifaces==0.11.0
notebook==6.4.8
ntlm-auth==1.4.0
numpy==1.21.5
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parsedatetime==2.6
parso==0.8.1
pbr==5.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
please-cli==0.1.1
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.28
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
py==1.10.0
py7zr==0.11.3+dfsg
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyAudio==0.2.11
pycairo==1.20.1
pychm==0.8.6
pycodestyle==2.8.0
pycryptodomex==3.11.0
pycups==2.0.1
pyee==8.2.2
pyfiglet==0.8.post1
Pygments==2.11.2
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pykerberos==1.1.14
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyppeteer==0.2.6
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
pyRFC3339==1.1
pyrsistent==0.18.1
pysmi==0.3.2
pysnmp==4.4.12
pystache==0.6.0
pyte==0.8.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.19.2
python-jsonstore==1.3.0
python-Levenshtein==0.12.2
pythran==0.10.0
pytz==2022.1
pywinrm==0.3.0
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.3.0
ranger-fm==1.9.3
regex==2021.11.10
reportlab==3.6.8
repoze.lru==0.7
requests==2.28.0
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rich==12.4.4
Routes==2.5.1
rsa==4.8
scipy==1.8.0
SecretStorage==3.3.1
selinux==3.3
Send2Trash==1.8.1b0
sgmllib3k==1.0.0
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
ssh-import-id==5.11
sympy==1.9
systemd-python==234
TatSu==5.8.0
terminado==0.13.1
testpath==0.5.0
texttable==1.6.4
thefuck==3.32
toml==0.10.2
tornado==6.1
tqdm==4.62.2
traitlets==5.1.1
tweepy==4.10.0
typer==0.4.1
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
ulauncher==5.14.4
unattended-upgrades==0.1
unicodedata2==14.0.0
uritemplate==3.0.1
urllib3==1.26.5
variety==0.8.5
vboxapi==1.0
vobject==0.9.6.1
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.6
websocket-client==1.2.3
websockets==9.1
widgetsnbextension==2.0.0
xdg==5
xkit==0.0.0
xmltodict==0.12.0
youtube-dl==2021.12.17
zeroconf==0.38.3
zim==0.74.3
zipp==1.0.0
